+++ plugins/hic/src/HicRenderer/index.ts
@@ -1,3 +1,17 @@
-export { default as ReactComponent } from './components/HicRendering'
-export { default as configSchema } from './configSchema'
-export { default } from './HicRenderer'
+import PluginManager from '@jbrowse/core/PluginManager'
+
+import ReactComponent from './components/HicRendering'
+import configSchema from './configSchema'
+import HicRenderer from './HicRenderer'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addRendererType(
+    () =>
+      new HicRenderer({
+        name: 'HicRenderer',
+        ReactComponent,
+        configSchema,
+        pluginManager,
+      }),
+  )
+}
