+++ plugins/hic/src/HicTrack/configSchema.ts
