+++ plugins/hic/src/HicTrack/index.ts
@@ -1,0 +1,15 @@
+import PluginManager from '@jbrowse/core/PluginManager'
+import TrackType from '@jbrowse/core/pluggableElementTypes/TrackType'
+import { createBaseTrackModel } from '@jbrowse/core/pluggableElementTypes/models'
+import configSchemaF from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addTrackType(() => {
+    const configSchema = configSchemaF(pluginManager)
+    return new TrackType({
+      name: 'HicTrack',
+      configSchema,
+      stateModel: createBaseTrackModel(pluginManager, 'HicTrack', configSchema),
+    })
+  })
+}
