+++ plugins/hic/src/LinearHicDisplay/index.ts
@@ -1,2 +1,20 @@
-export { default as configSchemaFactory } from './configSchema'
-export { default as modelFactory } from './model'
+import { DisplayType } from '@jbrowse/core/pluggableElementTypes'
+import PluginManager from '@jbrowse/core/PluginManager'
+import { BaseLinearDisplayComponent } from '@jbrowse/plugin-linear-genome-view'
+
+import configSchemaFactory from './configSchema'
+import stateModelFactory from './model'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addDisplayType(() => {
+    const configSchema = configSchemaFactory(pluginManager)
+    return new DisplayType({
+      name: 'LinearHicDisplay',
+      configSchema,
+      stateModel: stateModelFactory(configSchema),
+      trackType: 'HicTrack',
+      viewType: 'LinearGenomeView',
+      ReactComponent: BaseLinearDisplayComponent,
+    })
+  })
+}
